solving error

+++ demo/src/main/java/com/example/demo/repositories/LocalBusinessOwnerRepository.java
@@ -3,6 +3,6 @@ package com.example.demo.repositories;
 import com.example.demo.models.LocalBusinessOwner;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface LocalBusinessOwnerRepository extends JpaRepository<LocalBusinessOwner, Long> {
+public interface LocalBusinessOwnerRepository extends JpaRepository<LocalBusinessOwner, Integer> {
     
 }
